+++ Demo/API_V2/Assets/WX-WASM-SDK-V2/Editor/WXConvertCore.cs
@@ -18,34 +18,29 @@ namespace WeChatWASM
 
         static WXConvertCore()
         {
-            Init();
+            // Init();
         }
 
         public static void Init()
         {
-            config = UnityUtil.GetEditorConf();
-            // SDKFilePath = Path.Combine(Application.dataPath, "WX-WASM-SDK-V2", "Runtime", "wechat-default", "unity-sdk", "index.js");
             SDKFilePath = Path.Combine(UnityUtil.GetWxSDKRootPath(), "Runtime", "wechat-default", "unity-sdk", "index.js");
-            // string templateHeader = (UnityUtil.GetSDKMode() == UnityUtil.SDKMode.Package && UnityUtil.GetEngineVersion() >= UnityUtil.EngineVersion.Tuanjie) ? "PACKAGE:com.qq.wx.minigame:" : "PROJECT:";
             string templateHeader = "PROJECT:";
-#if PLATFORM_WEIXINMINIGAME
+#if TUANJIE_2022_3_OR_NEWER
             PlayerSettings.WeixinMiniGame.threadsSupport = false;
             PlayerSettings.runInBackground = false;
             PlayerSettings.WeixinMiniGame.compressionFormat = WeixinMiniGameCompressionFormat.Disabled;
-#if UNITY_2022_3_OR_NEWER
-            PlayerSettings.WeixinMiniGame.template = $"{templateHeader}WXTemplate2022TJ";
-#elif UNITY_2020_1_OR_NEWER
-            PlayerSettings.WeixinMiniGame.template = $"{templateHeader}WXTemplate2020";
-#else
-            PlayerSettings.WeixinMiniGame.template = $"{templateHeader}WXTemplate";
-#endif
+            if(UnityUtil.GetEngineVersion() == UnityUtil.EngineVersion.Tuanjie)
+            {
+                var absolutePath = Path.GetFullPath("Packages/com.qq.weixin.minigame/WebGLTemplates/WXTemplate2022TJ");
+                PlayerSettings.WeixinMiniGame.template = $"PATH:{absolutePath}";
+            }
+            else
+            {
+                PlayerSettings.WeixinMiniGame.template = $"{templateHeader}WXTemplate2022TJ";
+            }
             PlayerSettings.WeixinMiniGame.linkerTarget = WeixinMiniGameLinkerTarget.Wasm;
             PlayerSettings.WeixinMiniGame.dataCaching = false;
-#if UNITY_2021_2_OR_NEWER
             PlayerSettings.WeixinMiniGame.debugSymbolMode = WeixinMiniGameDebugSymbolMode.External;
-#else
-            PlayerSettings.WeixinMiniGame.debugSymbols = true;
-#endif
 #else
             PlayerSettings.WebGL.threadsSupport = false;
             PlayerSettings.runInBackground = false;
@@ -75,16 +70,27 @@ namespace WeChatWASM
             BUILD_WEBGL_FAILED = 2,
         }
 
-        public static WXEditorScriptObject config;
+        public static WXEditorScriptObject config => UnityUtil.GetEditorConf();
         public static string webglDir = "webgl"; // 导出的webgl目录
         public static string miniGameDir = "minigame"; // 生成小游戏的目录
         public static string audioDir = "Assets"; // 音频资源目录
+        public static string frameworkDir = "framework";
         public static string dataFileSize = string.Empty;
         public static string codeMd5 = string.Empty;
         public static string dataMd5 = string.Empty;
         private static string SDKFilePath = string.Empty;
         public static string defaultImgSrc = "Assets/WX-WASM-SDK-V2/Runtime/wechat-default/images/background.jpg";
-
+        public static bool UseIL2CPP
+        {
+            get
+            {
+#if TUANJIE_2022_3_OR_NEWER
+                return PlayerSettings.GetScriptingBackend(BuildTargetGroup.WeixinMiniGame) == ScriptingImplementation.IL2CPP;
+#else
+                return true;
+#endif
+            }
+        }
         // 可以调用这个来集成
         public static WXExportError DoExport(bool buildWebGL = true)
         {
@@ -122,10 +128,26 @@ namespace WeChatWASM
                 if (WXExtEnvDef.GETDEF("UNITY_2021_2_OR_NEWER") && !config.CompileOptions.DevelopBuild)
                 {
                     // 如果是2021版本，官方symbols产生有BUG，这里需要用工具将函数名提取出来
-                    WeChatWASM.UnityUtil.preprocessSymbols(GetWebGLSymbolPath());
+                    var symFile1 = "";
+                    if(!UseIL2CPP)
+                    {
+                        symFile1 = Path.Combine(config.ProjectConf.DST, webglDir, "Code", "wwwroot", "_framework", "dotnet.native.js.symbols");
+                    }
+                    else
+                    {
+                        var rootPath = Directory.GetParent(Application.dataPath).FullName;
+                        string webglDir = WXExtEnvDef.GETDEF("WEIXINMINIGAME") ? "WeixinMiniGame" : "WebGL";
+                        symFile1 = Path.Combine(rootPath, "Library", "Bee", "artifacts", webglDir, "build", "debug_WebGL_wasm", "build.js.symbols");
+                    }
+                    WeChatWASM.UnityUtil.preprocessSymbols(symFile1, GetWebGLSymbolPath());
+                    // WeChatWASM.UnityUtil.preprocessSymbols(GetWebGLSymbolPath());
                 }
 
                 ConvertCode();
+                if (!UseIL2CPP)
+                {
+                    ConvertDotnetCode();
+                }
                 string dataFilePath = GetWebGLDataPath();
                 string wxTextDataDir = WXAssetsTextTools.GetTextMinDataDir();
                 string dataFilePathBackupDir = $"{wxTextDataDir}{WXAssetsTextTools.DS}slim";
@@ -140,13 +162,13 @@ namespace WeChatWASM
                 }
                 File.Copy(dataFilePath, dataFilePathBackupPath);
 
-                if (config.CompileOptions.fbslim && !IsInstantGameAutoStreaming())
+                if (UseIL2CPP && config.CompileOptions.fbslim && !IsInstantGameAutoStreaming())
                 {
                     WXAssetsTextTools.FirstBundleSlim(dataFilePath, (result, info) =>
                     {
                         if (!result)
                         {
-                            Debug.LogError("【首包资源优化异常】：" + info);
+                            Debug.LogWarning("[首资源包跳过优化]：因处理失败自动跳过" + info);
                         }
 
                         finishExport();
@@ -248,15 +270,257 @@ namespace WeChatWASM
             return output.ToString();
         }
 
+        
+        private static void ConvertDotnetCode()
+        {
+            CompressAssemblyBrotli();
+            ConvertDotnetRuntimeCode();
+            ConvertDotnetFrameworkCode();
+        }
+
+        private static void ConvertDotnetRuntimeCode()
+        {
+            var runtimePath = GetWeixinMiniGameFilePath("jsModuleRuntime")[0];
+            var dotnetJs = File.ReadAllText(runtimePath, Encoding.UTF8);
+
+            Rule[] rules =
+            {
+                new Rule()
+                {
+                    old = "await *WebAssembly\\.instantiate\\(\\w*,",
+                    newStr = $"await WebAssembly.instantiate(Module[\"wasmPath\"],",
+                },
+                new Rule()
+                {
+                    old = "['\"]Expected methodFullName if trace is instrumented['\"]\\);?",
+                    newStr = "'Expected methodFullName if trace is instrumented'); return;",
+                }
+            };
+            foreach (var rule in rules)
+            {
+                if (ShowMatchFailedWarning(dotnetJs, rule.old, "runtime") == false) {
+                    dotnetJs = Regex.Replace(dotnetJs, rule.old, rule.newStr);
+                }
+            }
+
+            File.WriteAllText(Path.Combine(config.ProjectConf.DST, miniGameDir, frameworkDir, Path.GetFileName(runtimePath)), dotnetJs, new UTF8Encoding(false));
+        }
+
+        private static void CompressAssemblyBrotli()
+        {
+            GetWeixinMiniGameFilePath("assembly").ToList().ForEach(assembly => UnityUtil.brotli(assembly, assembly + ".br", "8"));
+        }
+
+        private static void ConvertDotnetFrameworkCode()
+        {
+            var target = "webgl.wasm.framework.unityweb.js";
+            var dotnetJsPath =
+                Path.Combine(config.ProjectConf.DST, webglDir, "Code", "wwwroot", "_framework", "dotnet.js");
+            var dotnetJs = File.ReadAllText(dotnetJsPath, Encoding.UTF8);
+            // todo: handle dotnet js
+            Rule[] rules =
+                {
+                    new Rule()
+                    {
+                        old = "import\\.meta\\.url",
+                        newStr = $"pathToFileURL('/{frameworkDir}/webgl.wasm.framework.unityweb.js')",
+                    },
+                    new Rule()
+                    {
+                        old = " *(\\w*) *= *(\\w*)\\(\"js-module-runtime\"\\) *, *(\\w*) *= *(\\w*)\\(\"js-module-native\"\\)",
+                        newStr = $" $1 = $2(\"js-module-runtime\"), $3 = $4(\"js-module-native\"); $1.resolvedUrl = \"{Path.GetFileName(GetWeixinMiniGameFilePath("jsModuleRuntime")[0])}\";$3.resolvedUrl = \"{Path.GetFileName(GetWeixinMiniGameFilePath("jsModuleNative")[0])}\";",
+                    },
+                    new Rule()
+                    {
+                        old = "export *\\{ *(.*) *as *default *,(.*) *as *dotnet *,(.*) *as *exit *\\} *;",
+                        newStr = "return {legacyEntrypoint: $1, dotnet: $2, exit: $3}",
+                    },
+                    new Rule()
+                    { // add symbol property for monoToBlazorAssetTypeMap
+                        old = "\"js-module-runtime\": *\"dotnetjs\", *\"js-module-threads\": *\"dotnetjs\"",
+                        newStr = "\"js-module-runtime\": \"dotnetjs\", \"js-module-threads\": \"dotnetjs\", \"symbols\": \"symbols\"",
+                    },
+                    new Rule()
+                    {
+                        old = "_e *\\|\\| *\"function\" *== *typeof *globalThis.URL[\\s\\S]*asm-features\"\\);",
+                        newStr = ""
+                    },
+                    new Rule()
+                    {
+                        old = "if *\\(!\\(ENVIRONMENT_IS_SHELL *\\|\\| *typeof *globalThis.URL *=== *\"function\"\\)\\)[\\s\\S]*BigInt64Array *API. *Please *use *a *modern *version. *See *also *https:\\/\\/aka.ms\\/dotnet\\-wasm\\-features\"\\);",
+                        newStr = ""
+                    }
+                };
+            foreach (var rule in rules)
+            {
+                if (ShowMatchFailedWarning(dotnetJs, rule.old, "dotnet") == false) {
+                    dotnetJs = Regex.Replace(dotnetJs, rule.old, rule.newStr);
+                }
+            }
+            var header = @"WebAssembly.validate = () => true;
+const dotnet = (function () {";
+            var footer = @"})();
+
+function pathToFileURL(path) {
+    return `file://${path}`;
+}
+
+const wxFetchFile = async (url, config) => new Promise((resolve, reject) => {
+    const folderPath = `${GameGlobal.manager.PLUGIN_CACHE_PATH}/${config.hash}`;
+    const filePath = `${folderPath}/${config.filename}`;
+    const fs = wx.getFileSystemManager();
+    var readFile = fs.readFile;
+    const readFileInfo = {
+        filePath: filePath,
+        success: async (res) => {
+            resolve(res.data);
+        },
+        fail: (err) => {
+            reject(err);
+        }
+    }
+    if (config.brotli) {
+        readFile = fs.readCompressedFile;
+        readFileInfo.compressionAlgorithm = 'br';
+    }
+    const downloadFileInfo = {
+        url: url,
+        success: (res) => {
+            if (res.statusCode === 200) {
+                try {
+                    fs.accessSync(folderPath);
+                } catch (e) {
+                    fs.mkdirSync(folderPath, true);
+                }
+                saveFile(res.tempFilePath, filePath, res.dataLength);
+            } else {
+                reject(res);
+            }
+        },
+        fail: (err) => {
+            reject(err);
+        }
+    };
+    const saveFile = (tempFilePath, filePath, dataLength, retry = 0) => {
+        try {
+            fs.saveFileSync(tempFilePath, filePath);
+            readFile(readFileInfo);
+        } catch (e) {
+            if (e.message === 'saveFileSync:fail exceeded the maximum size of the file storage limit') {
+                if (retry < 3) {
+                    GameGlobal.manager.cleanCache(dataLength).then(() => {
+                        saveFile(tempFilePath, filePath, dataLength, ++retry);
+                    });
+                } else {
+                    GameGlobal.manager.cleanAllCache().then(() => {
+                        wx.downloadFile(downloadFileInfo);
+                    });
+                }
+            } else {
+                reject(e);
+            }
+        }
+    };
+    // if not use cache or cache not exists, download file
+    fs.access({
+        path: filePath,
+        success: (res) => {
+            if (config.cache && res.errMsg === 'access:ok') {
+                readFile(readFileInfo);
+            } else {
+                wx.downloadFile(downloadFileInfo);
+            }
+        },
+        fail: (err) => {
+            wx.downloadFile(downloadFileInfo);
+        }
+    });
+});
+
+dotnet.dotnet.withResourceLoader((resourceType, assetName, url, requestHash, behavior) => {
+    const remoteUrl = `${GameGlobal.unityNamespace.DATA_CDN}Code/wwwroot/_framework/${assetName}`;
+    switch (resourceType) {
+        case 'dotnetjs':
+            return assetName;
+        case 'manifest':
+            return new Promise((resolve, reject) => {
+                wx.request({
+                    url: remoteUrl,
+                    success: (res) => {
+                        resolve({
+                            json: async () => res.data,
+                            headers: {
+                                get: () => undefined
+                            }
+                        });
+                    },
+                    fail: (err) => {
+                        reject(err);
+                    }
+                });
+            });
+        case 'symbols':
+        case 'dotnetwasm':
+            return new Promise((resolve, reject) => {
+                resolve({
+                    arrayBuffer: async () => undefined,
+                    status: 200,
+                    ok: true
+                })
+            });
+        case 'assembly':
+            const config = {
+                filename: assetName + '.br',
+                cache: true,
+                hash: requestHash,
+                brotli: true
+            };
+            return wxFetchFile(remoteUrl + '.br', config).then(res => ({
+                arrayBuffer: async () => res,
+                status: 200,
+                ok: true,
+            }));
+        default:
+            return url;
+    }
+});
+var executed = false;
+var unityFramework = function (module) {
+    if (executed) return;
+    executed = true;
+    module['noInitialRun'] = true;
+    return (dotnet.legacyEntrypoint(module)).then(() => {
+        module['callMain']();
+    });
+};
+if (typeof exports === 'object' && typeof module === 'object') module.exports = unityFramework;
+else if (typeof define === 'function' && define['amd']) define([], function () {
+    return unityFramework;
+});
+else if (typeof exports === 'object') exports['unityFramework'] = unityFramework;
+GameGlobal.unityNamespace.UnityModule = unityFramework;";
+            File.WriteAllText(Path.Combine(config.ProjectConf.DST, miniGameDir, frameworkDir, target), header + dotnetJs + footer, new UTF8Encoding(false));
+        }
+
         private static void ConvertCode()
         {
-            UnityEngine.Debug.LogFormat("[Converter] Starting to adapt framewor. Dst: " + config.ProjectConf.DST);
+            UnityEngine.Debug.LogFormat("[Converter] Starting to adapt framework. Dst: " + config.ProjectConf.DST);
 
             UnityUtil.DelectDir(Path.Combine(config.ProjectConf.DST, miniGameDir));
             string text = String.Empty;
+            var target = "webgl.wasm.framework.unityweb.js";
             if (WXExtEnvDef.GETDEF("UNITY_2020_1_OR_NEWER"))
             {
-                text = File.ReadAllText(Path.Combine(config.ProjectConf.DST, webglDir, "Build", "webgl.framework.js"), Encoding.UTF8);
+                if (UseIL2CPP)
+                {
+                    text = File.ReadAllText(Path.Combine(config.ProjectConf.DST, webglDir, "Build", "webgl.framework.js"), Encoding.UTF8);
+                }
+                else
+                {
+                    var frameworkPath = GetWeixinMiniGameFilePath("jsModuleNative")[0];
+                    target = Path.GetFileName(frameworkPath);
+                    text = File.ReadAllText(frameworkPath, Encoding.UTF8);
+                }
             }
             else
             {
@@ -266,7 +530,10 @@ namespace WeChatWASM
             for (i = 0; i < ReplaceRules.rules.Length; i++)
             {
                 var rule = ReplaceRules.rules[i];
-                text = Regex.Replace(text, rule.old, rule.newStr);
+                // text = Regex.Replace(text, rule.old, rule.newStr);
+                if (ShowMatchFailedWarning(text, rule.old, "WXReplaceRules") == false) {
+                    text = Regex.Replace(text, rule.old, rule.newStr);
+                }
             }
            string[] prefixs =
             {
@@ -318,7 +585,7 @@ namespace WeChatWASM
             {
                 text += ";GameGlobal.unityNamespace.UnityModule = tuanjieFramework;";
             }
-            else
+            else if (UseIL2CPP)
             {
                 if (text.StartsWith("(") && text.EndsWith(")"))
                 {
@@ -333,8 +600,13 @@ namespace WeChatWASM
                 Directory.CreateDirectory(Path.Combine(config.ProjectConf.DST, miniGameDir));
             }
 
-            var header = "function createWebAudio(){window.AudioContext=window.AudioContext||window.webkitAudioContext;if(window.AudioContext){return new AudioContext();}return wx.createWebAudioContext();}\n";
+            if (!Directory.Exists(Path.Combine(config.ProjectConf.DST, miniGameDir, frameworkDir)))
+            {
+                Directory.CreateDirectory(Path.Combine(config.ProjectConf.DST, miniGameDir, frameworkDir));
+            }
 
+            var header = "var OriginalAudioContext = window.AudioContext || window.webkitAudioContext;window.AudioContext = function() {if (this instanceof window.AudioContext) {return wx.createWebAudioContext();} else {return new OriginalAudioContext();}};";
+            
             if (config.CompileOptions.DevelopBuild)
             {
                 header = header + RenderAnalysisRules.header;
@@ -347,7 +619,52 @@ namespace WeChatWASM
 
             text = header + text;
 
-            File.WriteAllText(Path.Combine(config.ProjectConf.DST, miniGameDir, "webgl.wasm.framework.unityweb.js"), text, new UTF8Encoding(false));
+            var targetPath = Path.Combine(config.ProjectConf.DST, miniGameDir, target);
+            if (!UseIL2CPP)
+            {
+                targetPath = Path.Combine(config.ProjectConf.DST, miniGameDir, frameworkDir, target);
+                Rule[] nativeRules =
+                {
+                    new Rule()
+                    {
+                        old = "if\\(Module\\.IsWxGame\\)return Math\\.random\\(\\)\\*256\\|0;abort\\(\"randomDevice\"\\)",
+                        newStr = "{if(Module.IsWxGame)return Math.random()*256|0;abort(\"randomDevice\")}"
+                    },
+                    new Rule()
+                    {
+                        old = "var *_scriptDir *= *import\\.meta\\.url;",
+                        newStr = $"var _scriptDir = \"file:///framework/webgl.wasm.framework.unityweb.js\"",
+                    },
+                    new Rule()
+                    {
+                        old = "import\\.meta\\.url",
+                        newStr = "_scriptDir"
+                    },
+                    new Rule()
+                    {
+                        old = "Module\\[['\"](HEAP[U,F,1-9]*)['\"]\\] *=",
+                        newStr = "Module['$1'] = GameGlobal.unityNamespace.Module['$1'] =",
+                    },
+                    new Rule()
+                    {
+                        old = "return *handleException\\(e\\);?",
+                        newStr = "return handleException(e);} finally {if (ABORT === true) return; if (Module.calledMainCb) Module.calledMainCb(); if (GameGlobal.unityNamespace.enableProfileStats) {setTimeout(() => {SendMessage('WXSDKManagerHandler', 'OpenProfileStats');}, 10000);}"
+                    },
+                    new Rule()
+                    {
+                        old = "function *callMain\\(args *= *\\[\\]\\)",
+                        newStr = "Object.keys(Module).forEach(key=>{if(!(key in Object.keys(GameGlobal.unityNamespace.Module))){GameGlobal.unityNamespace.Module[key]=Module[key];}});function callMain(args = [])"
+                    },
+                };
+                foreach (var rule in nativeRules)
+                {
+                    if (ShowMatchFailedWarning(text, rule.old, "native") == false) {
+                        text = Regex.Replace(text, rule.old, rule.newStr);
+                    } 
+                }
+            }
+
+            File.WriteAllText(targetPath, text, new UTF8Encoding(false));
 
             UnityEngine.Debug.LogFormat("[Converter]  adapt framework done! ");
         }
@@ -358,14 +675,15 @@ namespace WeChatWASM
             PlayerSettings.WeixinMiniGame.emscriptenArgs = string.Empty;
             if (WXExtEnvDef.GETDEF("UNITY_2021_2_OR_NEWER"))
             {
-                PlayerSettings.WeixinMiniGame.emscriptenArgs += " -s EXPORTED_FUNCTIONS=_main,_sbrk,_emscripten_stack_get_base,_emscripten_stack_get_end";
+                // PlayerSettings.WeixinMiniGame.emscriptenArgs += " -s EXPORTED_FUNCTIONS=_main,_sbrk,_emscripten_stack_get_base,_emscripten_stack_get_end";
+                PlayerSettings.WeixinMiniGame.emscriptenArgs += " -s EXPORTED_FUNCTIONS=_sbrk,_emscripten_stack_get_base,_emscripten_stack_get_end -s ERROR_ON_UNDEFINED_SYMBOLS=0";
             }
 
 #else
             PlayerSettings.WebGL.emscriptenArgs = string.Empty;
             if (WXExtEnvDef.GETDEF("UNITY_2021_2_OR_NEWER"))
             {
-                PlayerSettings.WebGL.emscriptenArgs += " -s EXPORTED_FUNCTIONS=_sbrk,_emscripten_stack_get_base,_emscripten_stack_get_end";
+                PlayerSettings.WebGL.emscriptenArgs += " -s EXPORTED_FUNCTIONS=_sbrk,_emscripten_stack_get_base,_emscripten_stack_get_end -s ERROR_ON_UNDEFINED_SYMBOLS=0";
 #if UNITY_2021_2_5
                     PlayerSettings.WebGL.emscriptenArgs += ",_main";
 #endif
@@ -516,6 +834,13 @@ namespace WeChatWASM
             {
                 return Path.Combine(config.ProjectConf.DST, webglDir, "Build", "webgl.data.unityweb");
             }
+        }
+
+        private static string[] GetWeixinMiniGameFilePath(string key)
+        {
+            var bootJson = Path.Combine(config.ProjectConf.DST, webglDir, "Code", "wwwroot", "_framework", "blazor.boot.json");
+            var boot = JsonMapper.ToObject(File.ReadAllText(bootJson, Encoding.UTF8));
+            return boot["resources"][key].Keys.Select(file => Path.Combine(config.ProjectConf.DST, webglDir, "Code", "wwwroot", "_framework", file)).ToArray();
         }
 
         private static void finishExport()
@@ -692,7 +1017,14 @@ namespace WeChatWASM
         {
             if (WXExtEnvDef.GETDEF("UNITY_2020_1_OR_NEWER"))
             {
-                return Path.Combine(config.ProjectConf.DST, webglDir, "Build", "webgl.wasm");
+                if (UseIL2CPP)
+                {
+                    return Path.Combine(config.ProjectConf.DST, webglDir, "Build", "webgl.wasm");
+                }
+                else
+                {
+                    return GetWeixinMiniGameFilePath("wasmNative")[0];
+                }
             }
             else
             {
@@ -1073,6 +1405,8 @@ namespace WeChatWASM
                 IsInstantGameAutoStreaming() ? "true" : "false",
                 (config.CompileOptions.DevelopBuild && config.CompileOptions.enableRenderAnalysis) ? "true" : "false",
                 config.ProjectConf.IOSDevicePixelRatio.ToString(),
+                UseIL2CPP ? "" : "/framework",
+                UseIL2CPP ? "false" : "true",
             });
 
             List<Rule> replaceList = new List<Rule>(replaceArrayList);
@@ -1237,6 +1571,15 @@ namespace WeChatWASM
 #else
             return "";
 #endif
+        }
+
+        public static bool ShowMatchFailedWarning(string text, string rule, string file)
+        {
+            if (Regex.IsMatch(text, rule) == false) {
+                Debug.LogWarning($"UnMatched {file} rule: {rule}");
+                return true;
+            }
+            return false;
         }
     }