+++ Demo/API_V2/Assets/WX-WASM-SDK-V2/Editor/WXExtDef.cs
@@ -1,5 +1,6 @@
 ﻿using System.Collections;
 using System.Collections.Generic;
+using System.IO;
 using UnityEngine;
 using UnityEditor;
 using UnityEditor.Build;
@@ -105,6 +106,40 @@ namespace WeChatWASM
             // #endif
             //                 return null;
             //             });
+            WXExtEnvDef.RegisterAction("WXConvertCore.UseIL2CPP", (args) =>
+            {
+                return WXConvertCore.UseIL2CPP;
+            });
+            WXExtEnvDef.RegisterAction("UnityUtil.GetWxSDKRootPath", (args) => {
+#if UNITY_2018
+                return Path.Combine(Application.dataPath, "WX-WASM-SDK-V2");
+#else
+                var packageInfo = UnityEditor.PackageManager.PackageInfo.FindForAssembly(typeof(WXExtEnvDef).Assembly);
+                if (packageInfo == null)
+                {
+                    return Path.Combine(Application.dataPath, "WX-WASM-SDK-V2");
+                }
+                string packagePath = packageInfo.assetPath;
+                if (packageInfo.name == "WXSDK")
+                {
+                    packagePath += "/Resources";
+                }
+                DirectoryInfo dir = new DirectoryInfo(packagePath);
+                return dir.FullName;
+#endif
+            });
+            WXExtEnvDef.RegisterAction("UnityUtil.IsAssets", (args) => {
+#if UNITY_2018
+                return true;
+#else
+                var packageInfo = UnityEditor.PackageManager.PackageInfo.FindForAssembly(typeof(WXExtEnvDef).Assembly);
+                if (packageInfo == null)
+                {
+                    return true;
+                }
+                return false;
+#endif
+            });
         }
     }
 }